Implement the new cmake API
The main goal is to not force the user to LD_PRELOAD the probes. Another
feature is that the user does not have to deal with generated files in
the CMakeLists.txt anymore, everything is hidden in a qibuild submodule.

The user still has to #incule some generated files though.

the implementation is not fully complete:

 * the probe may currently be SHARED (ie LD_PRELOADed) or BUILTIN,
   support of STATIC is lacking/not tested
 * changing the probe linkage is a global option. Finer grained
   tweaking might be needed
 * WITH_PROBE is a global option, idem.
 * it would be great to generate some config file with all the probes an
   instrumented binary has to LD_PRELOAD. Not sure how to do that
   though. Add a command? Or do it inside qi_create_bin?

the legacy cmake API is still supported. The plan is to drop this support
as soon as we are happy with the new one.

Change-Id: Ia842c260b7bbaf0a2e27bcbcc2c20cc0a65a9018
Reviewed-on: http://gerrit.aldebaran.lan:8080/6443
Reviewed-by: sbarthelemy <sbarthelemy@aldebaran-robotics.com>
Tested-by: sbarthelemy <sbarthelemy@aldebaran-robotics.com>